backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
--no-binary psycopg2

asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
beautifulsoup4==4.9.3
    # via django-livereload-server
bokeh==2.3.3
    # via -r requirements.in
django==3.1.13
    # via
    #   -r requirements.in
    #   django-livereload-server
    #   django-revproxy
    #   djangorestframework
django-livereload-server==0.3.2
    # via -r requirements.in
django-revproxy==0.10.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
execnet==1.9.0
    # via pytest-xdist
importlib-metadata==4.6.4
    # via
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via bokeh
markupsafe==2.0.1
    # via jinja2
numpy==1.19.5
    # via
    #   bokeh
    #   pandas
packaging==21.0
    # via
    #   bokeh
    #   pytest
pandas==1.1.5
    # via -r requirements.in
pillow==8.3.2
    # via bokeh
pluggy==0.13.1
    # via pytest
psycopg2==2.9.1
    # via -r requirements.in
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
pytest-django==4.4.0
    # via -r requirements.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-xdist==2.3.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   bokeh
    #   pandas
pytz==2021.1
    # via
    #   django
    #   pandas
pyyaml==5.4.1
    # via bokeh
six==1.16.0
    # via
    #   django-livereload-server
    #   python-dateutil
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
tornado==6.1
    # via
    #   bokeh
    #   django-livereload-server
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   bokeh
    #   importlib-metadata
urllib3==1.26.6
    # via django-revproxy
zipp==3.5.0
    # via importlib-metadata
